+++ Sequences.h
@@ -16,15 +16,15 @@ public:
 	//getters
 	std::vector< std::vector<std::string> > getSequences();
 	std::vector< std::vector< std::vector<std::string> > > getEncodedSequences();
-	std::vector<std::string> performMSA(Profile*,int,std::string);
-	std::vector<std::string> performMSAencoded(std::vector<std::vector<double> >*,std::vector<std::vector<double> >*,int,double,std::string, bool);
-	std::vector<std::string> performMSAnextRound(Profile*,FeaturesProfile*,int,double,std::string, bool, double);
+	std::vector<std::string> performMSA(Profile*,double,std::string);
+	std::vector<std::string> performMSAencoded(std::vector<std::vector<double> >*,std::vector<std::vector<double> >*,double,double,std::string, bool);
+	std::vector<std::string> performMSAnextRound(Profile*,FeaturesProfile*,double,double,std::string, bool, double);
 private:
 	//functions
 	void removeGaps(std::string *,std::string *,std::vector<std::string> &);
 	void removeGaps(std::vector<std::string> *,std::vector<std::string> *,std::vector<std::vector<std::string> >&);
-	void alignPairwise(std::string *,std::string *, std::string, Profile&, int, int,std::string);
-	void alignPairwise(std::vector<std::string> *,std::vector<std::string> *, std::vector<std::string>, Profile&, FeaturesProfile&, int,double, int,std::string);
+	void alignPairwise(std::string *,std::string *, std::string, Profile&, double, int,std::string);
+	void alignPairwise(std::vector<std::string> *,std::vector<std::string> *, std::vector<std::string>, Profile&, FeaturesProfile&, double,double, int,std::string);
 	std::vector< std::vector<std::string> > sequences;	
 	std::vector< std::vector< std::vector<std::string> > > sequencesEncoded;	
 	double calcIdentity(const std::string&);
